// client/src/utilities/types/enums.ts
export enum MenuOrScheduleEnum {
  MENU = 'menu',
  SCHEDULE = 'schedule',
}

export enum ErrorVariantsEnum {
  NOT_FOUND = 'notFound',
  ERROR = 'error',
  NO_MENU = 'noMenu',
  NO_SCHEDULE = 'noSchedule',
}

export enum TextDirEnum {
  RTL = 'rtl',
  LTR = 'ltr',
}

export enum ApiResStatusEnum {
  LOADING = 'loading',
  SUCCESS = 'success',
  ERROR = 'error',
}

export enum DayOrWeekEnum {
  DAY = 'day',
  WEEK = 'week',
}